backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
anthropic>=0.9.0
pydantic>=2.4.2
httpx>=0.25.0
sqlalchemy>=2.0.22
python-multipart>=0.0.6 